4.20: middleware.js -- Added req, res and next variables to the get json function which now adds a token key to the request object and returns the next function app.js -- Added the middleware for token extraction to all routes blogs.js -- added the jwt package and implemented it in the post route handler
4.21:

blogs.js -- Adde authentication checks to the delte method handler. If there is no blog post or the decoded jwt user id does not match the id provided by the DB it will return a 401 code

+++ part_4/bloglist/server/controllers/blogs.js
@@ -2,15 +2,14 @@ const blogsRouter = require("express").Router();
 const logger = require("../utils/logger");
 const Blog = require("../models/blog");
 const User = require("../models/user");
-const middleware = require('../utils/middleware');
-const jwt = require('jsonwebtoken');
+const jwt = require("jsonwebtoken");
 
 blogsRouter.get("/", async (req, res) => {
   logger.info("Getting all entries");
   const blogs = await Blog.find({}, (err, result) => {
     if (err) return err;
     return result;
-  }).populate("user", { username: 1, name: 1, id: 1});
+  }).populate("user", { username: 1, name: 1, id: 1 });
   return blogs
     ? res.status(200).json(blogs)
     : res.status(404).json({ error: "No blogs found" });
@@ -27,11 +26,11 @@ blogsRouter.get("/:id", async (req, res) => {
 blogsRouter.post("/", async (req, res) => {
   logger.info("Creating new entry");
 
-  const token = middleware.getJwtToken(req);
+  const token = req.token;
   const decodedToken = jwt.verify(token, process.env.SECRET);
 
-  if ( !token || !decodedToken.id ) {
-    return res.status(401).json({ error: "Token missing or invalid" })
+  if (!token || !decodedToken.id) {
+    return res.status(401).json({ error: "Token missing or invalid" });
   }
 
   if (typeof req.body.title === "undefined") {
@@ -52,7 +51,7 @@ blogsRouter.post("/", async (req, res) => {
     author: req.body.author,
     url: req.body.url,
     user: user.id,
-  }
+  };
 
   const blog = new Blog(blogObject);
   const newBlog = await blog.save();
@@ -66,9 +65,29 @@ blogsRouter.post("/", async (req, res) => {
 });
 
 blogsRouter.delete("/:id", async (req, res) => {
+  const token = req.token;
+  const decodedToken = jwt.verify(token, process.env.SECRET);
+
+  if (!token || !decodedToken.id) {
+    return res.status(401).json({ error: "Token missing or invalid" });
+  }
+
+  const blog = await Blog.findById(req.params.id);
+
+  if ( !blog ) {
+    return res.status(401).json({ error: "No blog post found with the provided ID"})
+  }
+
+
+  if ( blog.user.toString() !== decodedToken.id.toString() ) {
+    return res.status(401).json({error: "Permission denied. You do not have the permission to delete this post."})
+  }
+
   logger.info("Deleting one entry");
   const response = await Blog.deleteOne({ _id: req.params.id });
-  res.status(200).json(response);
+  return response.deletedCount === 1
+    ? res.status(200).json({ message: "Deleted 1 blog" })
+    : res.status(401).json({ error: "There was a problem with the request" });
 });
 
 blogsRouter.put("/:id", async (req, res) => {
